refactor(edges/resolve): convert to a directory; lift the test module to tests.rs
index/edges/resolve.rs was 1351 lines (god-module #1 after the query_api split).
It's one cohesive job (edge resolution), so rather than split the logic, lift the
592-line #[cfg(test)] module into resolve/tests.rs. Production drops to 760 lines,
single-job and under the god-module threshold. Pure move — no behavior change.
